Move table formatting code to new file
